Added versioning
The versioning has been activated as follow:
v1.0: to keep it simple, the root "docs" directory will be the active version.
Due to Docusaurus built-in behaviors, the "docusaurus redirect plugin" has been added and configured.

v0.3: this is NOT the real v0.3 and instead a copy of the current v1.0 for the versioning purpose