Removing old information about install.networks
This statement is no longer true as of v.1.1, users can configure network interfaces for the host machine via install.management_interface. 

"Each key-value pair represents a network interface. The key name becomes the network name, and the values are configurations for each network."

This is done with install.networks in <v.1.1.