Change wording for network requirements from recommendation to required
For VM migration, for storage replication and for VM access in production scenarios / scenarios under load - 1 GB will not be sufficient at all. Basically with 1 GB the VM migration will not complete if the VM is under load and the automatic and working VM migration will also be required for Harvester upgrades while VMs are up and running.

So we really should change "10 G recommended" to "10 G required".