Update instructions for setting default GRUB boot menu entry
I've fixed this for v1.3 (dev) and v1.2 (latest) docs, but I've not
made any changes for v1.1 or earlier.  Judging from the source
for harvester-installer for these versions, we have a different file
to look at ("grubmenu" as opposed to "grubcustom"), because that's
what the older versions of elemental used when creating grub config.

Related issue: https://github.com/harvester/harvester/issues/4808

Signed-off-by: Tim Serong <tserong@suse.com>