docs: provide API flows for some operations
Provide API flows for uploading a VM image, creating a VM and
registering a Harvester cluster in a Rancher instance.
There are existing instructions on how to achieve the same via the UI,
however documented API flows make it possible for users to automate the
same actions.
These flows can serve as simple examples and users should find it fairly
easy to expand them to suite their needs, using the API reference.

related-to: https://github.com/harvester/harvester/issues/1816

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>