Versioned Sidebars: Add API sidebar links
Add API link definitions for the versioned sidebars for v1.1 and v1.2
API docs. This is used by docusaurus to generate the landing page for
the v1.2 and v1.1 API docs. Having this landing page is necessary to be
able to navigate to the individual pages of the v1.1 and v1.2 API docs
from the regular versioned docs.

fixes: https://github.com/harvester/harvester/issues/5948

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>