Add caution/info boxes for VM image virtual size
When creating volumes from images, it's important the the volume size
be greater than or equal to the size of the underlying image.  If this
is not the case, you'll potentially end up with a corrupt or truncated
filesystem in the volume.  This is particularly likely to occur with
qcow2 images, which can have a virtual size larger than the physical
image size.

In Harvester v1.3 and earlier, when creating VMs, the volume size
just defaults to 10GiB, regardless of the size of the underyling image.
When creating volumes on the volume screen, the size has no default.
In both cases it's hard to know the correct size to use, because
Harvester prior to v1.4 doesn't expose the virtual size of qcow2
images anywhere.  Because of this, I've used "caution" boxes for
the v1.3 docs.

In Harvester v1.4, the dashboard will default to setting the size of
volumes so they're at least as large as the virtual size of the image,
which means users are unlikely to run into trouble (unless they
deliberately choose to make the volume size smaller for some reason),
so I've dialed it back a bit and used the "info important" style for
this version.

If this looks generally OK, I'll copy the v1.3 docs back to v1.2
as well.

Related issue: https://github.com/harvester/harvester/issues/4905

Signed-off-by: Tim Serong <tserong@suse.com>