feat: LHv2: add note about non-NVMe disk size requirements
This one is a little obscure.  When testing with non-NVMe disks,
Longhorn uses the aio bdev driver, which assumes the block size
is 4096 bytes, and thus requires the disk to be an even multiple
of that figure.  We hit this in testing as described in
https://github.com/harvester/harvester/issues/6861.

Ordinarily one should be using NVMe disks, which end up using the
nvme bdev driver, which picks up the actual block size of the
underlying device and everything just works.  This subtlety
(nvme bdev driver working regardless of disk size, but aio bdev
driver requiring 4096 byte blocks) is apparently not documented
anywhere else (e.g. it's not mentioned in the Longhorn docs)
which is why I'm adding it here, just in case anyone else hits
it and gets confused.

Signed-off-by: Tim Serong <tserong@suse.com>