Mention LHv2 doesn't support trim with AIO bdev driver
Non-NVMe disks use the SPDK AIO bdev driver, which doesn't support
the unmap operation.  This means attempting to use trimming (also
known as discard) with these devices will result in I/O errors and
will pause your virtual machines.

This PR adds details about this limitation and gives an example for
how to work around it when creating ext4 filesystems in Linux VMs,
and also mentions how to disable trim for NTFS in Windows VMs.
For reference, the latter command was found in this forum post:

  https://www.tenforums.com/tutorials/40028-enable-disable-trim-support-solid-state-drives-windows-10-a.html

I've added this information to the existing note about size
limitations of non-NVMe disks, but if this note is getting too
long I guess we could always break it out into a separate section.

As with the size limitation, this issue only affects non-NVMe
disks. NVMe disks using the SPDK NVMe bdev driver support unmap
and thus trim/discard works just fine on those.

Signed-off-by: Tim Serong <tserong@suse.com>