v1.5 docs are mentioning v1.4 erroneously
There are several places where the documentation is mentioning v1.4 erroneously although it concerns v1.5.

Additionally there are several places that are mentioning v1.4.0 in the text which is redundant because it is already mentioned above in a `Available as of v1.4.0` note.

Fix the version from `1.4.0` to `1.4` because the doc relates to ALL versions not only to the patch version `.0`.

Related to: https://github.com/harvester/harvester/issues/7893

Signed-off-by: Volker Theile <vtheile@suse.com>