Add known issue for longhorn oversized volumes in v1.4.3
Harvester v1.4.3 updates Longhorn to v1.7.3, which has a problem with
oversized volumes being marked "Not Ready" and unable to be deleted.
I've added a note about this to the upgrade docs.

This issue is also present in Harvester v1.5.0, and has already been
documented, but that workaround refers to "/webhooks/0/rules/17", which
doesn't work for LH v1.7.3 (it's rule 14 in that case), so I've updated
the workaround to be a bit more generic and not depend on a specific
rule index.

Signed-off-by: Tim Serong <tserong@suse.com>