Revert to referencing kubectl section
Revert references to point to the FAQ section on how to take
the kubeconfig of harvester. Discussing how to explain it
made me think that the new section should be focused on
how to check the cluster state, not how to get kubeconfig as
we already have such section.

In case we change the place of the section or the contents
we need to only fix the reference, and it would be easier that
way as opposed to finding where we explained how kubeconfig
can be taken so we fix explanation with reference.

Signed-off-by: Martin Dekov <martin.dekov@suse.com>